Moved rates.py to util folder
