+++ examples/scripts/HestonSimulation.py
@@ -36,7 +36,7 @@ from quantlib.models.equity.heston_model import HestonModel
 from quantlib.quotes import SimpleQuote
 from quantlib.settings import Settings
 from quantlib.time.api import today, NullCalendar, ActualActual
-from quantlib.mlab.rates import flat_rate
+from quantlib.util.rates import flat_rate
 import pylab as pl
 from quantlib.sim.simulate import simulateHeston, simulateBates
